add wave_utils.hpp
